Update bodhi-appcenter.desktop
Improved the Dutch translations a bit. You might want to reconsider the existence of the nl_BE locale, by the way: it's the first instance of it that I've ever seen, and I've been translating into Dutch for over a decade for many projects. The Netherlands and Belgium are even united in an official Language Union (Taalunie) in order to keep Dutch unified. So the nl locale should suffice for Dutch, I think....